+++ src/main/java/edu/illinois/library/cantaloupe/resource/InformationRequestHandler.java
@@ -51,7 +51,7 @@ public class InformationRequestHandler extends AbstractRequestHandler
      */
     public static final class Builder {
 
-        private InformationRequestHandler handler;
+        private final InformationRequestHandler handler;
 
         private Builder(InformationRequestHandler handler) {
             this.handler = handler;
@@ -148,7 +148,14 @@ public class InformationRequestHandler extends AbstractRequestHandler
         boolean authorize() throws Exception;
 
         /**
-         * Called when the list of available output formats have been obtained
+         * Called immediately after a source image has been accessed.
+         *
+         * @param result Information about the source image.
+         */
+        void sourceAccessed(StatResult result);
+
+        /**
+         * Called when a list of available output formats has been obtained
          * from a {@link Processor}.
          */
         void knowAvailableOutputFormats(Set<Format> availableOutputFormats);
@@ -163,6 +170,9 @@ public class InformationRequestHandler extends AbstractRequestHandler
         @Override
         public boolean authorize() {
             return true;
+        }
+        @Override
+        public void sourceAccessed(StatResult sourceAvailable) {
         }
         @Override
         public void knowAvailableOutputFormats(Set<Format> availableOutputFormats) {
@@ -247,6 +257,7 @@ public class InformationRequestHandler extends AbstractRequestHandler
         if (optSrcImage.isEmpty() || isResolvingFirst()) {
             try {
                 StatResult result = source.stat();
+                callback.sourceAccessed(result);
             } catch (NoSuchFileException e) { // this needs to be rethrown!
                 if (config.getBoolean(Key.CACHE_SERVER_PURGE_MISSING, false)) {
                     // If the image was not found, purge it from the cache.