+++ src/main/java/edu/illinois/library/cantaloupe/resource/iiif/v1/InformationResource.java
@@ -1,11 +1,7 @@
 package edu.illinois.library.cantaloupe.resource.iiif.v1;
 
-import java.time.Instant;
-import java.time.ZoneId;
-import java.time.format.DateTimeFormatter;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Locale;
 import java.util.Set;
 
 import edu.illinois.library.cantaloupe.http.Method;
@@ -15,6 +11,7 @@ import edu.illinois.library.cantaloupe.resource.JacksonRepresentation;
 import edu.illinois.library.cantaloupe.resource.ResourceException;
 import edu.illinois.library.cantaloupe.resource.Route;
 import edu.illinois.library.cantaloupe.resource.InformationRequestHandler;
+import edu.illinois.library.cantaloupe.source.StatResult;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -66,6 +63,14 @@ public class InformationResource extends IIIF1Resource {
                 }
                 return false;
             }
+
+            @Override
+            public void sourceAccessed(StatResult result) {
+                if (result.getLastModified() != null) {
+                    setLastModifiedHeader(result.getLastModified());
+                }
+            }
+
             @Override
             public void knowAvailableOutputFormats(Set<Format> formats) {
                 availableOutputFormats.addAll(formats);
@@ -103,11 +108,7 @@ public class InformationResource extends IIIF1Resource {
                 String.format("<%s>;rel=\"profile\";", iiifInfo.profile));
         // Last-Modified
         if (info.getSerializationTimestamp() != null) {
-            getResponse().setHeader("Last-Modified",
-                    DateTimeFormatter.RFC_1123_DATE_TIME
-                            .withLocale(Locale.UK)
-                            .withZone(ZoneId.systemDefault())
-                            .format(info.getSerializationTimestamp()));
+            setLastModifiedHeader(info.getSerializationTimestamp());
         }
     }