+++ src/test/java/edu/illinois/library/cantaloupe/resource/iiif/v3/InformationResourceTest.java
@@ -531,7 +531,7 @@ public class InformationResourceTest extends ResourceTest {
         client = newClient("/" + IMAGE + "/info.json");
         Response response = client.send();
         Headers headers = response.getHeaders();
-        assertEquals(7, headers.size());
+        assertEquals(8, headers.size());
 
         // Access-Control-Allow-Origin
         assertEquals("*", headers.getFirstValue("Access-Control-Allow-Origin"));
@@ -542,8 +542,8 @@ public class InformationResourceTest extends ResourceTest {
                 headers.getFirstValue("Content-Type")));
         // Date
         assertNotNull(headers.getFirstValue("Date"));
-        // Last-Modified (only present for cached infos, tested separately)
-        assertNull(headers.getFirstValue("Last-Modified"));
+        // Last-Modified
+        assertNotNull(headers.getFirstValue("Last-Modified"));
         // Server
         assertNotNull(headers.getFirstValue("Server"));
         // Vary